+++ benchmarks/dynamo/common.py
@@ -23,6 +23,7 @@ import psutil
 import torch
 
 import torch._dynamo
+import torch._dynamo.backends.inductor
 import torch._dynamo.utils
 import torch.distributed
 from scipy.stats import gmean, ttest_ind
@@ -1911,7 +1912,7 @@ def run(runner, args, original_dir=None):
     args.exclude = args.exclude or [r"^$"]
     args.exclude_exact = args.exclude_exact or []
 
-    if args.inductor:
+    if torch._dynamo.backends.inductor.inductor:
         assert args.backend is None
         args.backend = "inductor"
     if args.dynamic_ci_skips_only:
@@ -2067,7 +2068,7 @@ def run(runner, args, original_dir=None):
     if args.devices == ["cpu"]:
         runner.skip_models.update(runner.very_slow_models)
 
-    if args.inductor or args.inductor_settings:
+    if torch._dynamo.backends.inductor.inductor or args.inductor_settings:
         runner.skip_models.update(runner.failing_torchinductor_models)
         if args.float16:
             # TODO(jansel): check if correctness issue is real
@@ -2097,7 +2098,7 @@ def run(runner, args, original_dir=None):
         optimize_ctx = torch._dynamo.optimize(dummy_fx_compile, nopython=args.nopython)
         experiment = speedup_experiment
         output_filename = "overheads.csv"
-    elif args.inductor:
+    elif torch._dynamo.backends.inductor.inductor:
         inductor_config.debug = args.verbose
         if args.threads:
             inductor_config.cpp.threads = args.threads
@@ -2180,7 +2181,7 @@ def run(runner, args, original_dir=None):
         experiment = coverage_experiment
         output_filename = "coverage.csv"
 
-    if args.inductor or args.backend == "inductor":
+    if torch._dynamo.backends.inductor.inductor or args.backend == "inductor":
         if args.disable_cudagraphs:
             inductor_config.triton.cudagraphs = False
 
@@ -2208,7 +2209,7 @@ def run(runner, args, original_dir=None):
         if args.profiler_trace_name is None:
             if args.backend:
                 args.profiler_trace_name = args.backend
-            elif args.inductor:
+            elif torch._dynamo.backends.inductor.inductor:
                 args.profiler_trace_name = "inductor"
             else:
                 args.profiler_trace_name = "profile"
