+++ torch/_dynamo/backends/registry.py
@@ -81,9 +81,7 @@ def _lazy_import():
 
     # TODO(jansel): refactor backends defined in other places
     from .. import debug_utils
-    from ..optimizations import backends, training
+    from ..optimizations import backends
 
-    training.create_aot_backends()
-    # avoid unused import lint
     assert backends is not None
     assert debug_utils is not None
