+++ torch/_inductor/compile_fx.py
@@ -14,10 +14,10 @@ import torch._dynamo.config as dynamo_config
 import torch.fx
 
 from torch._dynamo import logging as dynamo_logging, utils as dynamo_utils
-from torch._dynamo.optimizations.training import aot_autograd
 from torch._dynamo.utils import fake_mode_from_tensors
 from torch._functorch.aot_autograd import make_boxed_func
 from torch._subclasses.fake_tensor import FakeTensor
+from .._dynamo.backends.common import aot_autograd
 from . import config, metrics, overrides, pattern_matcher
 from .debug import DebugContext
 from .decomposition import select_decomp_table
